+++ src/action/input.ts
@@ -10,6 +10,7 @@ export class Input {
   private run_?: string
   private operatingSystem_?: os.Kind
   private version_?: string
+  private imageURL_?: string
   private shell_?: Shell
   private environmentVariables_?: string
   private architecture_?: architecture.Kind
@@ -26,6 +27,11 @@ export class Input {
     return (this.version_ = core.getInput('version', {
       required: true
     }))
+  }
+
+  get imageURL(): string {
+    if (this.imageURL_ !== undefined) return this.imageURL_
+    return (this.imageURL_ = core.getInput('image_url'))
   }
 
   get operatingSystem(): os.Kind {