+++ changelog.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add support for custom image URLs ([#13](https://github.com/cross-platform-actions/action/pull/13))
 
 ## [0.17.0] - 2023-07-25
 ### Changed